+++ src/reconfox/reporting/__init__.py
@@ -4,6 +4,7 @@ from reconfox.reporting.html import render_html
 from reconfox.reporting.json_report import render_json
 from reconfox.reporting.markdown import render_markdown
 from reconfox.reporting.ndjson import render_ndjson
+from reconfox.reporting.sarif import render_sarif
 from reconfox.reporting.writer import (
     ReportFormat,
     default_filename,
@@ -20,6 +21,7 @@ __all__ = [
     "render_json",
     "render_markdown",
     "render_ndjson",
+    "render_sarif",
     "write_report",
     "write_report_to_file",
 ]